+++ src/main/java/com/rocks/zipcode/service/dto/FoodMenuDTO.java
@@ -53,11 +53,11 @@ public class FoodMenuDTO implements Serializable {
         this.quantityOfItem = quantityOfItem;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
